requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.2.0
click==8.1.8
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==42.0.5
cssselect2==0.7.0
daphne==4.1.0
Django==4.2
django-appconf==1.0.6
django-crontab==0.7.1
django-dbbackup==4.2.1
django-autoslug==1.9.9
django-cors-headers==4.1.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-imagekit==5.0.0
django-jazzmin==2.6.0
django-tables2==2.7.0
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.2.0
et_xmlfile==2.0.0
ethiopian-date==1.0
exceptiongroup==1.2.2
fonttools==4.51.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
mysqlclient==2.2.0
numpy==1.24.4
opencv-python==4.9.0.80
openpyxl==3.1.5
packaging==24.0
pilkit==3.0
pillow==10.2.0
pure-python-adb==0.3.0.dev0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pydyf==0.9.0
Pygments==2.19.1
pyOpenSSL==24.1.0
pyphen==0.14.0
pytesseract==0.3.10
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2023.3
PyYAML==6.0.2
redis==5.0.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.41.3
tablib==3.6.1
tinycss2==1.2.1
Twisted==24.3.0
txaio==23.1.1
typer==0.15.1
tzdata==2023.4
urllib3==2.0.4
uvicorn==0.33.0
watchfiles==0.24.0
weasyprint==61.2
webencodings==0.5.1
websockets==13.1
zope.interface==6.2
zopfli==0.2.3
